Committed 2021/09/23

+++ Matplotlib 3 Cookbook/Stacked plot2.py
@@ -1,7 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-x = np.array([1, 2, 3, 4, 5, 6], dtype=np.int32)
+x = np.array([1, 2, 3, 4, 5, 6])
 April = [5, 7, 6, 7, 8, 9]
 May = [0, 4, 3, 7, 8, 9]
 June = [6, 7, 4, 5, 6, 8]
